% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{anual}
\alias{anual}
\title{Función para poner parametrización del formato anual}
\usage{
anual(
  color1 = rgb(0, 0, 1),
  color2 = rgb(0.615686274509804, 0.733333333333333, 1),
  fuente = pkg.env$fuente
)
}
\arguments{
\item{color1}{Color primario para graficar, definido por el usuario. Este debe ser definido 
con el formato rgb(v1,v2,v3, maxColorValue = 255), si maxColorValue no está definido
se usa por defecto 1.}

\item{color2}{Color secundario para graficar, definido por el usuario con el formato 
rgb(v1,v2,v3, maxColorValue = 255), si maxColorValue no está definido
se usa por defecto 1. 
rgb(0,0,1), rgb(0.6156862745098039,0.7333333333333333,1)}
}
\description{
Función para poner parametrización del formato anual
}
