% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabras.R
\name{buscarPalabraDiccionario}
\alias{buscarPalabraDiccionario}
\title{Función que busca una palabra en el diccionario}
\usage{
buscarPalabraDiccionario(key)
}
\arguments{
\item{key}{}
}
\description{
Función que busca una palabra en el diccionario
}
