% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculosLatex.R
\name{calcularAlto}
\alias{calcularAlto}
\title{Función que calcula el alto de un cuadrado de texto dentro de LaTeX}
\usage{
calcularAlto(texto, factor = 0.9, largo = 20)
}
\arguments{
\item{texto}{Es el texto que se desea poner dentro del rectángulo en LaTeX}

\item{factor}{Es el factor de escalamiento para el texto dentro del
código tikz}

\item{largo}{Es la longitud máxima para el cuadro de texto. Dicha longitud debe ser ingresada en milimetros.}
}
\value{
Alto del cuadro que enmarca al texto ingresado. Este valor es devuelto en puntos.
}
\description{
Función que calcula el alto de un cuadrado de texto dentro de LaTeX
}

