% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{calcularPosicionesDobles}
\alias{calcularPosicionesDobles}
\title{Funcion que calcula las posiciones para las etiquetas en las graficas de linea Doble}
\usage{
calcularPosicionesDobles(graph)
}
\arguments{
\item{graph}{Objeto del tipo ggplot2 al cual se le quiere poner las etiquetas}
}
\value{
Un vector indicando las posiciones de las etiquetas
}
\description{
Funcion que calcula las posiciones para las etiquetas en las graficas de linea Doble
}

