% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{calcularRampa}
\alias{calcularRampa}
\title{Hace una paleta del color indicado}
\usage{
calcularRampa(data, color1 = pkg.env$color1, color2 = pkg.env$color2)
}
\arguments{
\item{data}{El data frame con la informacion}

\item{color1}{El color primario con el cual se desea hacer la paleta}

\item{color2}{El color secundario que va cuando encuentra palabras clave}
}
\value{
Vector con la rampa de colores
}
\description{
Hace una paleta del color indicado
}

