% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{calcularRampaAnillo}
\alias{calcularRampaAnillo}
\title{Calcula la rampa de colores para usar en las graficas
de anillo, mandando los ignorados hasta el final y haciendo
la separcion por categorias (blanco para una categoría y negro para la otra)
El gris siempre se usa para ignorado}
\usage{
calcularRampaAnillo(x, categoria = TRUE)
}
\arguments{
\item{x}{El vector de datos en el cual se basa la paleta de colores}

\item{categoria}{Booleano que indica si se desea categorizar la rampa}
}
\value{
El vector de la paleta de colores
}
\description{
Calcula la rampa de colores para usar en las graficas
de anillo, mandando los ignorados hasta el final y haciendo
la separcion por categorias (blanco para una categoría y negro para la otra)
El gris siempre se usa para ignorado
}

