% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGramaticales.R
\name{calcularRespuestaPor}
\alias{calcularRespuestaPor}
\title{Función para comparación de porcentajes
La cota sirve para saber que mensaje se debe emitir,
la función está pensada en el documento de vitales}
\usage{
calcularRespuestaPor(data, cota = 1, primeraPos = 5, ultimaPos = 9)
}
\arguments{
\item{data}{Data frame con el que se desea trabajar.}

\item{cota}{Valor númerico que indica que cota se usará.}

\item{primeraPos}{Posición inicial del vector y en data frame.}

\item{ultimaPos}{Posición final del vector y en el data frame.}
}
\description{
Función para comparación de porcentajes
La cota sirve para saber que mensaje se debe emitir,
la función está pensada en el documento de vitales
}

