% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cambiarCodificacion}
\alias{cambiarCodificacion}
\title{Cambia la codificación en ciertas partes del data frame}
\usage{
cambiarCodificacion(tabla)
}
\arguments{
\item{tabla}{Es el data frame al cual se le desea cambiar la
codificacion}
}
\value{
Data frame recodificado
}
\description{
Cambia la codificación en ciertas partes del data frame
}

