% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cambioFormato.R
\name{cambiarGraficas}
\alias{cambiarGraficas}
\title{Funcion para cambiar los ambitos de la graficas}
\usage{
cambiarGraficas(tamFuente)
}
\description{
Funcion para cambiar los ambitos de la graficas
}

