% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabras.R
\name{cambiarNombres}
\alias{cambiarNombres}
\title{Función para cortar nombres, versión 1}
\usage{
cambiarNombres(data, noColumn)
}
\arguments{
\item{data}{El data frame con la información necesaria}

\item{noColumn}{El número de columna sobre el cúal se quiere hacer el análisis de la longitud de la palabra}
}
\description{
Función para cortar nombres, versión 1
}
