% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cambioInterAnual}
\alias{cambioInterAnual}
\title{Función que calcula el cambio interanual en porcentaje para un data frame dado}
\usage{
cambioInterAnual(data, primeraPos = 5, ultimaPos = 9)
}
\arguments{
\item{data}{El data frame sobre el cual se desea hacer el calculo}

\item{paso}{El paso de retroceso para el calculo}
}
\value{
Cambio interanual
}
\description{
Función que calcula el cambio interanual en porcentaje para un data frame dado
}

