% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cambioInterAnualNeto}
\alias{cambioInterAnualNeto}
\title{Función que calcula el cambio interanual neto para un data frame dado}
\usage{
cambioInterAnualNeto(data, paso = 4)
}
\arguments{
\item{data}{El data frame sobre el cual se desea hacer el calculo}

\item{paso}{El paso de retroceso para el calculo}
}
\value{
Cambio interanual
}
\description{
Función que calcula el cambio interanual neto para un data frame dado
}

