% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cargaMasiva}
\alias{cargaMasiva}
\title{Funci칩n para hacer carga masiva
de archivos CSV para su posterior uso con una lista
dentro de R.}
\usage{
cargaMasiva(ruta, recodificar = F)
}
\arguments{
\item{ruta}{Ruta dentro del disco duro en la cual est치n contenidos los CSV}
}
\value{
Una lista con los data frame que contiene la informaci칩n.
}
\description{
Funci칩n para hacer carga masiva
de archivos CSV para su posterior uso con una lista
dentro de R.
}

