% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{compilar}
\alias{compilar}
\title{Función que comila con xelatex y muestra el resultado}
\usage{
compilar(ruta = "")
}
\arguments{
\item{ruta}{Es la ruta de donde se desea guardar el fichero .tex}
}
\description{
Función que comila con xelatex y muestra el resultado
}

