% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{completarEtiquetas}
\alias{completarEtiquetas}
\title{Funcion interna que le pone el relleno al vector de etiquetas
del tipo ("","",etiqueta,"",..)}
\usage{
completarEtiquetas(dato, posicion, tam = 5)
}
\arguments{
\item{dato}{Dato que va en el vector}

\item{posicion}{Es la posicion que ocupa el dato dentro del vector}

\item{tam}{El tamano del vector de salida}
}
\value{
El vector completo rellenado de la forma ("", "", dato, "", ...)
}
\description{
Funcion interna que le pone el relleno al vector de etiquetas
del tipo ("","",etiqueta,"",..)
}

