% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{convertirFechasIPC}
\alias{convertirFechasIPC}
\title{Función leer convertirFechas de IPC}
\usage{
convertirFechasIPC(lista)
}
\value{
lista Una lista con los data frame que contiene la información.
}
\description{
Función leer convertirFechas de IPC
}
