% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cortarEtiquetas}
\alias{cortarEtiquetas}
\title{Función para cortar las etiquetas cuando son muy largas para las gráficas}
\usage{
cortarEtiquetas(etiquetas)
}
\arguments{
\item{etiquetas}{Es el vector que contiene las etiquetas del eje x o el eje y según sea el caso}
}
\description{
Función para cortar las etiquetas cuando son muy largas para las gráficas
}
