% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabras.R
\name{crearDiccionarioPalabras}
\alias{crearDiccionarioPalabras}
\title{Función que crea el diccionario en caso de que no esté}
\usage{
crearDiccionarioPalabras()
}
\description{
Función que crea el diccionario en caso de que no esté
}
