% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cuatroEtiquetas}
\alias{cuatroEtiquetas}
\title{Funcion para activar las cuatro etiquetas, maximo y minimo.}
\usage{
cuatroEtiquetas(encendido = TRUE)
}
\description{
Funcion para activar las cuatro etiquetas, maximo y minimo.
}
