% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{escribirCSV}
\alias{escribirCSV}
\title{Función para escribir CSV}
\usage{
escribirCSV(lista, ruta)
}
\arguments{
\item{lista}{Es la lista que contiene los data frame}
}
\description{
Función para escribir CSV
}
