% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{escribirLibro}
\alias{escribirLibro}
\title{Función para escribir un libro de excel a partir de una lista de data frames}
\usage{
escribirLibro(lista, ruta)
}
\description{
Función para escribir un libro de excel a partir de una lista de data frames
}
