% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasBarras}
\alias{etiquetasBarras}
\title{Anota las etiquetas para una grafica de barras}
\usage{
etiquetasBarras(graph, margenIz = 0, precision = 1, cambiarNegativas = F)
}
\arguments{
\item{graph}{Objeto ggplot2 que se desea anotar}
}
\value{
Retorna objeto ggplot2 listo para graficar

cambiarNegativas parámetro temporal, desaparecerá en futuras versiones.
}
\description{
Anota las etiquetas para una grafica de barras
}
