% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasBarras2}
\alias{etiquetasBarras2}
\title{Anota las etiquetas para una grafica de barras
Haciendo pruebas de no medir el largo de las etiquetas con 
TikzDevice y hacerlo puramente con ggplot2}
\usage{
etiquetasBarras2(graph, margenIz = 0, precision = 1, cambiarNegativas = F)
}
\arguments{
\item{graph}{Objeto ggplot2 que se desea anotar}
}
\value{
Retorna objeto ggplot2 listo para graficar

cambiarNegativas parámetro temporal, desaparecerá en futuras versiones.
}
\description{
Anota las etiquetas para una grafica de barras
Haciendo pruebas de no medir el largo de las etiquetas con 
TikzDevice y hacerlo puramente con ggplot2
}
