% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasFacets}
\alias{etiquetasFacets}
\title{Pone etiquetas a las columnas en una grafica de columnas}
\usage{
etiquetasFacets(graph, precision = 1)
}
\arguments{
\item{graph}{El objeto ggplot2 que se desea anotar}
}
\value{
Objeto ggplot2 anotado listo para usar
}
\description{
Pone etiquetas a las columnas en una grafica de columnas
}
