% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasLineas}
\alias{etiquetasLineas}
\title{Le pone las etiquetas a una grafica de linea}
\usage{
etiquetasLineas(graph, posiciones)
}
\arguments{
\item{graph}{Objeto del tipo ggplot2 que desea anotar}

\item{posiciones}{Vector de posiciones en que van las etiquetas}
}
\description{
Le pone las etiquetas a una grafica de linea
}

