% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasLineasDobles}
\alias{etiquetasLineasDobles}
\title{Le pone las etiquetas a una grafica de linea}
\usage{
etiquetasLineasDobles(graph, pos, precision = 1)
}
\arguments{
\item{graph}{Objeto del tipo ggplot2 que desea anotar}

\item{precision}{Numero de decimales con el que se desea el vector respuesta.
Por defecto se usa un decimal}

\item{posiciones}{Vector de posiciones en que van las etiquetas}
}
\value{
Gráfica con las etiquetas puestasw
}
\description{
Le pone las etiquetas a una grafica de linea
}

