% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{excluirNiveles}
\alias{excluirNiveles}
\title{Función para ordenar los niveles de un data frame
excluyendo ciertas palabras claves, como ignorado y otros.
Se puede hacer personalizable eligiendo los identificadores de los
niveles que se desean excluir.}
\usage{
excluirNiveles(data, palabras = pkg.env$exclusion)
}
\arguments{
\item{data}{El data frame con el que se desea trabajar}

\item{palabras}{Vector de palabras claves que se desean excluir,}
}
\value{
Data frame ordenado y con las exclusiones de los niveles
}
\description{
Función para ordenar los niveles de un data frame
excluyendo ciertas palabras claves, como ignorado y otros.
Se puede hacer personalizable eligiendo los identificadores de los
niveles que se desean excluir.
}

