% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{existeTraslape}
\alias{existeTraslape}
\title{Funcion en fase que beta, en teoria mide el ancho de una palabra y determina si existe el espacio
suficiente para que quepa en la grafica}
\usage{
existeTraslape(graph, ancho = 0.6)
}
\arguments{
\item{graph}{Objeto del tipo ggplot2 que se desea modificar}

\item{ancho}{El ancho de las barras en porcentaje, segun indicaciones de ggplot2}
}
\description{
Funcion en fase que beta, en teoria mide el ancho de una palabra y determina si existe el espacio
suficiente para que quepa en la grafica
}

