% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{exportarLatex}
\alias{exportarLatex}
\title{Exporta a codigo tikz en LaTeX usando tikzDevice}
\usage{
exportarLatex(nombre = grafica.tex, graph, preambulo = F)
}
\arguments{
\item{nombre}{Ruta del fichero LaTeX}

\item{graph}{Objeto ggplot2 que se desea exportar a LaTeX}
}
\description{
Exporta a codigo tikz en LaTeX usando tikzDevice
}

