% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{fact2Num}
\alias{fact2Num}
\title{Convierte los factores de un data frame a datos numericos}
\usage{
fact2Num(tabla)
}
\arguments{
\item{tabla}{El data frame que se quiere trabajar}
}
\value{
Regresa el data frame con valores numéricos en lugar de factores
}
\description{
Convierte los factores de un data frame a datos numericos
}

