% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaltasJudiciales.R
\name{getListJudiciales}
\alias{getListJudiciales}
\title{Función para obtener de nuevo la lista de datos para faltas judiciales}
\usage{
getListJudiciales()
}
\description{
Función para obtener de nuevo la lista de datos para faltas judiciales
}
