% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPC.R, R/cb.R
\name{getPath}
\alias{getPath}
\title{Funcion para obtener la ruta en la cual se exportan los csv}
\usage{
getPath()

getPath()
}
\description{
Funcion para obtener la ruta en la cual se exportan los csv

Funcion para obtener la ruta en la cual se exportan los csv
}
