% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaltasJudiciales.R
\name{getPathJudiciales}
\alias{getPathJudiciales}
\title{Funcion para obtener la ruta en la cual se exportan los tex con las gráficas}
\usage{
getPathJudiciales()
}
\description{
Funcion para obtener la ruta en la cual se exportan los tex con las gráficas
}
