% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficasEspeciales.R
\name{graficaBarFacets}
\alias{graficaBarFacets}
\title{Funci칩n para gr치ficar con facets.}
\usage{
graficaBarFacets(data, ruta, escala = "normal", etiquetas = "H", preambulo = F)
}
\arguments{
\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}

\item{Data}{Frame con el que se desea trabajar}
}
\description{
Funci칩n para gr치ficar con facets.
}
