% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaCol}
\alias{graficaCol}
\title{Funcion basica para graficar Columnas}
\usage{
graficaCol(data, color1 = pkg.env$color1, ancho = 0.6, ordenar = TRUE,
  escala = "normal")
}
\arguments{
\item{data}{Data frame para elaborar la grafica}

\item{color1}{Color con el que se desea hacer la grafica}

\item{ancho}{Porcentaje que ocupan las columnas, segun paquete ggplot2}

\item{ordenar}{Booleano que indica si los datos deben ser ordenados}

\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}

\item{color2}{Color secundario para las graficas}
}
\value{
La gráfica lista para agregarle etiquetas y toques finales
}
\description{
Funcion basica para graficar Columnas
}

