% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graficasEspeciales.R
\name{graficaColCategorias}
\alias{graficaColCategorias}
\title{Graficas de columnas por categorias}
\usage{
graficaColCategorias(data, etiquetasCategorias = "A", escala = "normal",
  ruta, etiquetas = "v", ancho = 0.8, preambulo = F, ejeX = "h")
}
\arguments{
\item{data}{El data frame para hacer la grafica, con el formato de tres o mas columnas de la forma y,z,w,...}

\item{etiquetasCategorias}{Indica la posición en las que se desea poner las etiquetas para las categorias.
Por defecto la posición es hacia Arriba, lo que se denota con la letra "A", cuando las etiquetas se desean en la parte
derecha de la grafica se debe indicar con la letra "D"}

\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}

\item{ruta}{Nombre de la ruta en la que se desea guardar la salida del archivo tex para su compilacion}

\item{etiquetas}{Posición de las etiquetas. Por defecto se ponen verticales, para pasar a horizontales de escribir "h"}

\item{preambulo}{Etiqueta boolean que indica si se desea que la gráfica tenga preámbulo o no. Por defecto se tiene Falso.}
}
\value{
No regresa ningun valor
}
\description{
Graficas de columnas por categorias
}

