% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaDobleLinea}
\alias{graficaDobleLinea}
\title{Genera graficas de dobles lineas en un mismo panel.}
\usage{
graficaDobleLinea(data, ruta, preambulo = F, color1 = pkg.env$color1,
  color2 = pkg.env$color2, ancho = 1.5, precision = 1,
  escala = "normal", inicio = 0, fin = 0, ejeX = "h")
}
\arguments{
\item{data}{El data frame para hacer la gráfica}

\item{ruta}{Es la ruta en disco donde se desea almacenar la grafica}

\item{preambulo}{Booleano que indica si una grafica debe llevar preambulo o no.}

\item{color1}{El color en el que se desea la linea}

\item{ancho}{El grosor de la linea}

\item{precision}{Se refiere al número de decimales que se desean mostrar en la gráfica. Por defecto se usa un decimal.}

\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}

\item{inicio}{El dato desde donde se quiere que se visualice la gráfica}

\item{fin}{El dato hasta donde se desea visualizar la gráfica}

\item{ejeX}{Parámetro de tipo cadena que determina la orientación de las etiquetas para el eje X. Las opciones son "h" para horizontal y
"v" para vertical. Por defecto se usa horizontal.}
}
\description{
Genera graficas de dobles lineas en un mismo panel.
}

