% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaLinea}
\alias{graficaLinea}
\title{Hace una grafica de linea, util para series historicas}
\usage{
graficaLinea(data, color1 = pkg.env$color1, inicio = 0, ancho = 1.5)
}
\arguments{
\item{data}{El data frame para hacer la gráfica}

\item{color1}{El color en el que se desea la linea}

\item{inicio}{El dato desde donde se quiere que se visualice la gráfica}

\item{ancho}{El grosor de la linea}
}
\description{
Hace una grafica de linea, util para series historicas
}

