% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{graficasCanasta}
\alias{graficasCanasta}
\title{Funcion para generar las graficas de la canasta básica}
\usage{
graficasCanasta(modalidad = "trimestral")
}
\description{
Funcion para generar las graficas de la canasta básica
}
