% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio.r
\name{graficasComercio}
\alias{graficasComercio}
\title{Funci칩n para hacer la gr치ficas trimestrales
de comercio exterior}
\usage{
graficasComercio(lista, ruta, modalidad = "trimestral")
}
\arguments{
\item{lista}{Es el listado que contiene los data frame de comercio exterior}

\item{ruta}{Es el path de salida para los tex con las graficas}
}
\description{
Funci칩n para hacer la gr치ficas trimestrales
de comercio exterior
}
