% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabras.R
\name{guardarPalabraEnDiccionario}
\alias{guardarPalabraEnDiccionario}
\title{Función que guarda el resumen de una palabra larga cuando el usuario la ingresa}
\usage{
guardarPalabraEnDiccionario(key, corta)
}
\arguments{
\item{key}{}

\item{corta}{}
}
\description{
Función que guarda el resumen de una palabra larga cuando el usuario la ingresa
}
