% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{inc2mm}
\alias{inc2mm}
\title{Funcion para convertir de pulgadas a milimetros}
\usage{
inc2mm(number)
}
\arguments{
\item{number}{Dato a ser convertido}
}
\value{
Valor en milimetros
}
\description{
Funcion para convertir de pulgadas a milimetros
}

