% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Funcion experimental que determina cuando un dato es entero o no
solamente datos enteros o del tipo flotante}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Dato numerico}
}
\value{
Un valor booleano indicando si el dato es entero o flotante
}
\description{
Funcion experimental que determina cuando un dato es entero o no
solamente datos enteros o del tipo flotante
}

