% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{leerLibroNormal}
\alias{leerLibroNormal}
\title{Funci칩n leer libro excel}
\usage{
leerLibroNormal(ruta, codificacion = "iso")
}
\arguments{
\item{ruta}{Ruta dentro del disco duro en la cual est치n contenidos los CSV}
}
\value{
Una lista con los data frame que contiene la informaci칩n.
}
\description{
Funci칩n leer libro excel
}
