% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correo.R
\name{mandarCorreo}
\alias{mandarCorreo}
\title{Función para mandar los correos de reportes}
\usage{
mandarCorreo(direccion, asunto, cuerpo, ruta)
}
\arguments{
\item{direccion}{Es la dirección del destinatario}

\item{asunto}{Es el asunto del correo}
}
\description{
Función para mandar los correos de reportes
}
