% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{mm2inch}
\alias{mm2inch}
\title{Funcion para convertir de milimetros a pulgadas}
\usage{
mm2inch(number)
}
\arguments{
\item{number}{Dato a ser convertido}
}
\value{
Valor en pulgadas
}
\description{
Funcion para convertir de milimetros a pulgadas
}

