% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{mm2pt}
\alias{mm2pt}
\title{Funcion para convertir de milimetros a puntos}
\usage{
mm2pt(number)
}
\arguments{
\item{number}{Dato a ser convertido}
}
\value{
Valor en puntos
}
\description{
Funcion para convertir de milimetros a puntos
}

