% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{ordenarNiveles}
\alias{ordenarNiveles}
\title{Funcion de uso interno para ordenar los niveles según los datos}
\usage{
ordenarNiveles(data, ordenar = TRUE)
}
\arguments{
\item{data}{El data frame con el cual se desean hacer los calculos}

\item{ordenar}{Booleno que indica si se deben ordenar los datos de mayor a menor}
}
\value{
Vector con los valores en el orden deseado
}
\description{
Funcion de uso interno para ordenar los niveles según los datos
}

