% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{pdfcrop}
\alias{pdfcrop}
\title{Función que hace pdfcrop}
\usage{
pdfcrop(ruta = "")
}
\arguments{
\item{ruta}{Es la ruta de donde se desea guardar el fichero .tex}

\item{mostrar}{Booleano, cuando es verdadero muestra el pdf compilado.}
}
\description{
Función que hace pdfcrop
}
