% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{piramidePoblacional}
\alias{piramidePoblacional}
\title{Función para hacer gráficas piramidales de población}
\usage{
piramidePoblacional(
  data,
  ancho = 0.6,
  escala = "normal",
  color1 = pkg.env$color1,
  ruta,
  preambulo = F,
  digitos = 1,
  porcentual = T
)
}
\arguments{
\item{data}{Es el data frame con el que se hará la gráfica}

\item{ancho}{Gradua el ancho de las Barras. Por defecto es 0.6}

\item{color1}{El color con el que se desea hacer la gráfica}

\item{ruta}{La ruta completa de donde se desea almacenar la gráfica}

\item{digitos}{El número de dígitos que se usa para los decimales}
}
\description{
Función para hacer gráficas piramidales de población
}
