% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{posicionesBarras}
\alias{posicionesBarras}
\title{Función para calcular vector de posiciones para las etiquetas de Barras}
\usage{
posicionesBarras(x)
}
\arguments{
\item{x}{Vector de valores para el eje y.}
}
\description{
Función para calcular vector de posiciones para las etiquetas de Barras
}

