% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{presentacion}
\alias{presentacion}
\title{Funci贸n para poner parametrizaci贸n del formato de presentaciones}
\usage{
presentacion()
}
\description{
Funci贸n para poner parametrizaci贸n del formato de presentaciones
}
