% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{pt2mm}
\alias{pt2mm}
\title{Convierte de puntos a milimetros}
\usage{
pt2mm(unidad)
}
\arguments{
\item{unidad}{Valor numerico que desea ser convertido}
}
\value{
Valor en milimetros
}
\description{
Convierte de puntos a milimetros
}

