% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{rampaColAgrupadas}
\alias{rampaColAgrupadas}
\title{Calcula la rampa de colores para usar en las graficas
de columnas agrupadas, mandando los ignorados hasta el final y haciendo
la separcion por categorias (blanco para una categoría y negro para la otra)
El gris siempre se usa para ignorado}
\usage{
rampaColAgrupadas(data)
}
\arguments{
\item{data}{El data frame con el que se harán los calculos}
}
\value{
El vector de la paleta de colores
}
\description{
Calcula la rampa de colores para usar en las graficas
de columnas agrupadas, mandando los ignorados hasta el final y haciendo
la separcion por categorias (blanco para una categoría y negro para la otra)
El gris siempre se usa para ignorado
}

