% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{rotarEtiX}
\alias{rotarEtiX}
\title{Funcion que rota las etiquetas en el eje X para graficas de columnas que poseen
etiquetas horizontales para las columnas}
\usage{
rotarEtiX(graph)
}
\arguments{
\item{graph}{El objeto ggpot2 que se desea modificar}
}
\value{
Objeto ggplot2 modificado en las etiquetas del eje X
}
\description{
Funcion que rota las etiquetas en el eje X para graficas de columnas que poseen
etiquetas horizontales para las columnas
}

