% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{rotarEtiX2}
\alias{rotarEtiX2}
\title{Funcion que rota las etiquetas en el eje X para graficas de columnas que poseen
etiquetas verticales para las columnas}
\usage{
rotarEtiX2(graph)
}
\arguments{
\item{graph}{El objeto ggpot2 que se desea modificar}
}
\value{
Objeto ggplot2 modificado en las etiquetas del eje X
}
\description{
Funcion que rota las etiquetas en el eje X para graficas de columnas que poseen
etiquetas verticales para las columnas
}

