% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{setListCB}
\alias{setListCB}
\title{Funci칩n encargada de establecer la lista donde estar치n alojadados los datos
para la cb}
\usage{
setListCB(lista)
}
\description{
Funci칩n encargada de establecer la lista donde estar치n alojadados los datos
para la cb
}
