% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IPC.R
\name{setListIpc}
\alias{setListIpc}
\title{Funci칩n encargada de establecer la lista donde estar치n alojadados los datos
para el IPC}
\usage{
setListIpc(lista)
}
\description{
Funci칩n encargada de establecer la lista donde estar치n alojadados los datos
para el IPC
}

