% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaltasJudiciales.R
\name{setListJudiciales}
\alias{setListJudiciales}
\title{Función para setear la lista con los data frames para faltas judiciales}
\usage{
setListJudiciales(lista)
}
\description{
Función para setear la lista con los data frames para faltas judiciales
}
