% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPC.R, R/cb.R
\name{setPath}
\alias{setPath}
\title{Funcion para establecer la ruta en la cual se exportarán los csv}
\usage{
setPath(ruta)

setPath(ruta)
}
\arguments{
\item{ruta}{Es la ruta dentro del disco duro donde se almacenaran los csv}
}
\description{
Funcion para establecer la ruta en la cual se exportarán los csv

Funcion para establecer la ruta en la cual se exportarán los csv
}
