% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaltasJudiciales.R
\name{setPathJudiciales}
\alias{setPathJudiciales}
\title{Funcion para establecer la ruta en la cual se exportarán los tex con las graficas}
\usage{
setPathJudiciales(ruta)
}
\arguments{
\item{ruta}{Es la ruta dentro del disco duro donde se almacenaran los csv}
}
\description{
Funcion para establecer la ruta en la cual se exportarán los tex con las graficas
}
