% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{sonEnteros}
\alias{sonEnteros}
\title{Funcion experimental que determina si un vector numerico
esta conformado exclusivamente por valores entero o no}
\usage{
sonEnteros(data)
}
\arguments{
\item{Vector}{numerico}
}
\value{
Un valor booleano indicando si el vector es de enteros o no
}
\description{
Funcion experimental que determina si un vector numerico
esta conformado exclusivamente por valores entero o no
}

