% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{trimestral}
\alias{trimestral}
\title{Funci贸n para poner parametrizaci贸n del formato de trimestrales}
\usage{
trimestral()
}
\description{
Funci贸n para poner parametrizaci贸n del formato de trimestrales
}

