% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{anual}
\alias{anual}
\title{Función para poner parametrización del formato anual}
\usage{
anual(color1, color2)
}
\arguments{
\item{color1}{Color primario para graficar, definido por el usuario. Este debe ser definido
con el formato rgb(v1,v2,v3, maxColorValue = 255), si maxColorValue no está definido
se usa por defecto 1.}

\item{color2}{Color secundario para graficar, definido por el usuario con el formato
rgb(v1,v2,v3, maxColorValue = 255), si maxColorValue no está definido
se usa por defecto 1.}
}
\description{
Función para poner parametrización del formato anual
}

