% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{cuatroEtiquetas}
\alias{cuatroEtiquetas}
\title{Funcion para activar las cuatro etiquetas, maximo y minimo.}
\usage{
cuatroEtiquetas(encendido = T)
}
\description{
Funcion para activar las cuatro etiquetas, maximo y minimo.
}

