% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{etiquetasHorizontales}
\alias{etiquetasHorizontales}
\title{Pone etiquetas a las columnas en una grafica de columnas}
\usage{
etiquetasHorizontales(graph, precision = 1)
}
\arguments{
\item{graph}{El objeto ggplot2 que se desea anotar}
}
\value{
Objeto ggplot2 anotado listo para usar
}
\description{
Pone etiquetas a las columnas en una grafica de columnas
}

