% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaAnillo}
\alias{graficaAnillo}
\title{Hace graficas de anillo para dos y tres categorias}
\usage{
graficaAnillo(data, nombre)
}
\arguments{
\item{data}{Data frame con el cual se hace el anillo}

\item{nombre}{Ruta en la cual se desea generar el tex}
}
\value{
No retorna
}
\description{
Hace graficas de anillo para dos y tres categorias
}

