% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaCol}
\alias{graficaCol}
\title{Funcion basica para graficar Columnas}
\usage{
graficaCol(data, color1 = pkg.env$color1, ancho = 0.6, ordenar = TRUE)
}
\arguments{
\item{data}{Data frame para elaborar la grafica}

\item{color1}{Color con el que se desea hacer la grafica}

\item{ancho}{Porcentaje que ocupan las columnas, segun paquete ggplot2}

\item{ordenar}{Booleano que indica si los datos deben ser ordenados}

\item{color2}{Color secundario para las graficas}
}
\description{
Funcion basica para graficar Columnas
}

