% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaColCategorias}
\alias{graficaColCategorias}
\title{Graficas de columnas por categorias}
\usage{
graficaColCategorias(data)
}
\arguments{
\item{data}{El data frame para hacer la grafica, con el formato de tres o mas columnas de la forma y,z,w,...}
}
\description{
Graficas de columnas por categorias
}

