% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaColCategorias}
\alias{graficaColCategorias}
\title{Graficas de columnas por categorias}
\usage{
graficaColCategorias(data, etiquetasCategorias = "I", escala = "normal")
}
\arguments{
\item{data}{El data frame para hacer la grafica, con el formato de tres o mas columnas de la forma y,z,w,...}

\item{etiquetasCategorias}{Indica la posición en las que se desea poner las etiquetas para las categorias.
Por defecto la posición es a la izquierda, lo que se denota con la letra "I", cuando las etiquetas se desean en la parte
superior de la grafica se debe indicar con la letra "A"}

\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}
}
\value{
No regresa ningun valor
}
\description{
Graficas de columnas por categorias
}

