% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaLinea}
\alias{graficaLinea}
\title{Hace una grafica de linea, util para series historicas}
\usage{
graficaLinea(data, color1 = pkg.env$color1, inicio = 0, ancho = 1.5,
  precision = 1, escala = "normal", rotar = T)
}
\arguments{
\item{data}{El data frame para hacer la gráfica}

\item{color1}{El color en el que se desea la linea}

\item{inicio}{El dato desde donde se quiere que se visualice la gráfica}

\item{ancho}{El grosor de la linea}

\item{precision}{Se refiere al número de decimales que se desean mostrar en la gráfica. Por defecto se usa
un decimal.}

\item{escala}{Indica la escala en la cual debe estar el eje y de la grafica. Por defecto se encuentra en normal. Las opciones
son "miles", "millones" o "milesmillones".}
}
\description{
Hace una grafica de linea, util para series historicas
}

