% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesGraficas.R
\name{graficaLineaTrim}
\alias{graficaLineaTrim}
\title{Genera graficas de lineas para series historicas de los trimestrales}
\usage{
graficaLineaTrim(data, color1 = color, inicio = 0, ancho = 0.5)
}
\arguments{
\item{data}{El data frame con el que se hará la grafica}

\item{color1}{El color en el que se desea la linea}

\item{inicio}{El punto en el eje y a partir del cual se desea mostrar la grafica}

\item{El}{ancho de la linea}
}
\value{
El objeto ggplot2 listo para grafica
}
\description{
Genera graficas de lineas para series historicas de los trimestrales
}

