% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vitales.R
\name{graficasVitales}
\alias{graficasVitales}
\title{Función encargada de fabricar las graficas para
las estadísticas vitales.}
\usage{
graficasVitales(lista, ruta)
}
\arguments{
\item{lista}{Listado de R que contiene los data frame de vitales}

\item{ruta}{Ruta de salida para los archivos de las gráficas}
}
\description{
Función encargada de fabricar las graficas para
las estadísticas vitales.
}

