% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcionesINE.R
\name{inc2pt}
\alias{inc2pt}
\title{Funcion para convertir de pulgadas a puntos}
\usage{
inc2pt(number)
}
\arguments{
\item{number}{Dato a ser convertido}
}
\value{
Valor en puntos
}
\description{
Funcion para convertir de pulgadas a puntos
}

